Fix for 12.4.
With 12.3, the DSL was all method_missing, so methods from the superclass
would always take priority. Now that we are loading the DSL via Poise,
the DSL is ahead of the provider base methods, and the git provider
has a method named #git which conflicted. This basically patches around
the problem, but the long term solution is the DSL should be included
in Chef::Provider again.